// src/components/types.ts
// global file for resuable types throughout components

export type IconVector =
  | 'Feather'
  | 'Ionicons'
  | 'EvilIcons'
  | 'Entypo'
  | 'FontAwesome'
  | 'FontAwesome6'
  | 'FontAwesome5'
  | 'MaterialCommunityIcons'
  | 'MaterialIcons'
  | 'Fontisto'
  | 'Foundation'
  | 'FontAwesome5Pro'
  | 'FontAwesome6Pro'
  | 'Octicons'
  | 'AntDesign';